refactor(blorktools): rig handle factory
